Removed unnecessary casting

+++ MD02-715-Java-Track-Lecture-2/MD02-763-Task-3-Go-Fighters-Application/src/com/mentormate/devcamp/game/hero/Hero.java
@@ -125,7 +125,7 @@ public abstract class Hero {
      * @return the double
      */
     protected double randomizeValueByPercentage(long value) {
-        return (double) value * (random.nextDouble() * (MAX_PERCENTAGE - MIN_PERCENTAGE) + MIN_PERCENTAGE);
+        return value * (random.nextDouble() * (MAX_PERCENTAGE - MIN_PERCENTAGE) + MIN_PERCENTAGE);
     }
 
     /**
